Added basic support for sonars. In order to use them, set the private use_sonar parameter to True, and publish a transform (e.g. using the static_transform_publisher) from the centre of the sonar ring to the base_link (or other base frame) of the robot. This means that you basically only need to account for the z-offset of the sonars from the base_link.
This code has not been tested on robots with multiple sonar rings at different heights (such as the Peoplebot) and is likely to require some modifications: The ARIA library does not appear to return z coordinates of sonar readings (or sonar senors), which means that other means would be necessary to distinguish between the upper and lower sonar ring of a Peoplebot. Additionally, the difference in x,y,z of the sonar rings would need to be applied to the sonar readings of the upper ring so that the published point cloud is correct.