Add watchdog to stop robot if no cmd_vel received
Without this commit, the robot simply continues driving at the speed
from the last received `cmd_vel` message, even if the node that is
supposed to send the `cmd_vel` messages crashes. It is standard behavior
in ROS for robots to stop in this case (at least that's what the PR2 and
every other robot I have worked with so far does).